requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
toml==0.10.2
